requirements.txt
Django
djangorestframework
gunicorn
django-cors-headers